requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==23.2
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.13.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.5
webdriver-manager==4.0.1
wsproto==1.2.0
psycopg2-binary==2.9.11